Restore ability to copy-paste keys directly.
fbc7d4d7e2f7cbe04d4ae26152a21c5fc1a1d387 has introduced URL handling of
a "compressed" boolean per public key. The primary purpose is to
signal the tool that one of the keys is compressed. That works well when
the fields are pre-populated from the Coinbase url.

However this has broken another use-case of the recovery tool. Directly
inputting the public keys from the printed vault backup, or
copy-pasting them manually from the 'vault details' page on Coinbase
will yield incorrect addresses, as the code will make the assumption
that the coinbase key is uncompressed. This commit fixes it.

As the 3 keys are not interchangeable anymore, the fields are now named
and the keys need to be entered in the correct fields.